requirements.txt
fastapi==0.95.1
jinja2==3.1.2
uvicorn==0.21.1
uvloop==0.17.0
python-multipart==0.0.6
cassandra-driver==3.27.0
passlib==1.7.4
pyjwt==2.6.0
python-dotenv==1.0.0